Moved initialize_logger
Moves implementation of initialize_logger inside __main__ in order
to reduce internal module imports.
